for the most part, skeleton of slideshow is working; could be better but gonna move on for now
