+++ src/data/slides.ts
@@ -1,0 +1,29 @@
+import type { SlideProps } from '../types/types';
+
+export const slides: SlideProps[] = [
+    {
+        title: "PASAE",
+        description: "Pilipinx Association of Scientists, Architects, and Engineers",
+        imageSrc: "https://pasae.studentorg.berkeley.edu/images/front_page/PASAE37.JPG"
+    },
+    {
+        title: "Professionalism",
+        description: "Network, Mentors, Resumes, Interview Prep",
+        imageSrc: "https://pasae.studentorg.berkeley.edu/images/professionalism_fr.jpg"
+    },
+    {
+        title: "Funnovation",
+        description: "Have Fun!",
+        imageSrc: "https://pasae.studentorg.berkeley.edu/images/front_page/funnovation37.JPG"
+    },
+    {
+        title: "Family",
+        description: "A Community",
+        imageSrc: "https://pasae.studentorg.berkeley.edu/images/front_page/family37.jpg"
+    },
+    {
+        title: "Get Involved!",
+        description: "It's easy!",
+        imageSrc: "https://pasae.studentorg.berkeley.edu/images/front_page/get_involvedCORE37.jpg"
+    }
+];
