+++ src/types/types.tsx
@@ -1,0 +1,5 @@
+export interface SlideProps {
+    imageSrc: string;
+    title: string;
+    description: string;
+}
